Refactored using hooks

+++ src/containers/app.container.jsx
@@ -1,7 +1,8 @@
+import React, { useState, useEffect } from 'react';
+import "@babel/polyfill";
+
 import LinkInput from '../components/LinkInput';
 import ProgressBar from '../components/ProgressBar';
-
-import React, { Component } from 'react';
 
 import * as path from 'path';
 
@@ -12,56 +13,48 @@ const { ipcRenderer, remote } = window.require('electron');
 const ytdl = window.require('ytdl-core');
 const fs = window.require('fs-extra');
 
-class AppContainer extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      url: '',
-      showProgressBar: false,
-      progress: 0,
-      bitrate: localStorage.getItem('userBitrate')
-        ? parseInt(localStorage.getItem('userBitrate'))
-        : 160,
-      progressMessage: '',
-      userDownloadsFolder: localStorage.getItem('userSelectedFolder')
-        ? localStorage.getItem('userSelectedFolder')
-        : remote.app.getPath('downloads'),
-    };
+function AppContainer (props) {
+  const [url, setUrl] = useState('');
 
+  const [progress, setProgress] = useState(0);
+  const [showProgress, setShowProgress] = useState(false);
+  const [progressMessage, setProgressMessage] = useState('');
+
+  const [bitrate, setBitrate] = useState(
+    localStorage.getItem('userBitrate')
+      ? parseInt(localStorage.getItem('userBitrate'))
+      : 160
+  )
+
+  const [folder, setFolder] = useState(
+    localStorage.getItem('userSelectedFolder')
+      ? localStorage.getItem('userSelectedFolder')
+      : remote.app.getPath('downloads')
+  )
+
+  // This property will be used to control the rate at which the progress bar is updated to prevent UI lag.
+  let rateLimitTriggered = false;
+
+  useEffect(() => {
+    alert('effect');
     ipcRenderer.on('changeBitrate', (event, newBitrate) => {
-      this.setState({ bitrate: newBitrate });
+      setBitrate(newBitrate);
       localStorage.setItem('userBitrate', newBitrate.toString());
     });
 
     // Signal from main process to show prompt to change the download to folder.
     ipcRenderer.on('promptForChangeDownloadFolder', () => {
       // Changing the folder in renderer because we need access to both state and local storage.
-      this.changeOutputFolder();
+      changeOutputFolder();
     });
+  })
 
-    // This property will be used to control the rate at which the progress bar is updated to prevent UI lag.
-    this.rateLimitTriggered = false;
-
-    this.startDownload = this.startDownload.bind(this);
-    this.downloadFinished = this.downloadFinished.bind(this);
-    this.changeOutputFolder = this.changeOutputFolder.bind(this);
-    this.updateUrl = this.updateUrl.bind(this);
-  }
-
-  updateUrl(e) {
-    e.preventDefault();
-    this.setState({
-      url: e.target.value
-    });
-  }
-
-  getVideoAsMp4(urlLink, userProvidedPath, title) {
+  const getVideoAsMp4 = (urlLink, userProvidedPath, title) => {
     // Tell the user we are starting to get the video.
-    this.setState({ progressMessage: 'Downloading...' });
+    setProgressMessage('Downloading...');
     title = sanitize(title);
     return new Promise((resolve, reject) => {
       let fullPath = path.join(userProvidedPath, `tmp_${title}.mp4`);
-      // console.log('fullPath', fullPath)
 
       // Create a reference to the stream of the video being downloaded.
       let videoObject = ytdl(urlLink, {
@@ -70,19 +63,18 @@ class AppContainer extends Component {
       });
 
       videoObject.on('progress', (chunkLength, downloaded, total) => {
-        console.log('progress');
         // When the stream emits a progress event, we capture the currently downloaded amount and the total
         // to download, we then divided the downloaded by the total and multiply the result to get a float of
         // the percent complete, which is then passed through the Math.floor function to drop the decimals.
-        if (!this.rateLimitTriggered) {
+        if (!rateLimitTriggered) {
           let newVal = Math.floor((downloaded / total) * 100);
-          this.setState({ progress: newVal });
+          setProgress(newVal);
 
           // Set the rate limit trigger to true and set a timeout to set it back to false. This will prevent the UI
           // from updating every few milliseconds and creating visual lag.
-          this.rateLimitTriggered = true;
+          rateLimitTriggered = true;
           setTimeout(() => {
-            this.rateLimitTriggered = false;
+            rateLimitTriggered = false;
           }, 800);
         }
       });
@@ -91,7 +83,7 @@ class AppContainer extends Component {
       videoObject.pipe(fs.createWriteStream(fullPath)).on('finish', () => {
         // all of the video stream has finished piping, set the progress bar to 100% and give user pause to see the
         // completion of step. Then we return the path to the temp file, the output path, and the desired filename.
-        this.setState({ progress: 100 });
+        setProgress(100)
         setTimeout(() => {
           resolve({
             filePath: fullPath,
@@ -103,27 +95,27 @@ class AppContainer extends Component {
     });
   }
 
-  convertMp4ToMp3(paths) {
+  const convertMp4ToMp3 = (paths) => {
     console.log('convertMp4ToMp3');
     // Tell the user we are starting to convert the file to mp3.
-    this.setState({ progressMessage: 'Converting...', progress: 0 });
+    setProgress(0);
+    setProgressMessage('Converting...')
 
     return new Promise((resolve, reject) => {
       // Reset the rate limiting trigger just encase.
-      this.rateLimitTriggered = false;
+      rateLimitTriggered = false;
 
       // Pass ffmpeg the temp mp4 file. Set the path where is ffmpeg binary for the platform. Provided desired format.
       ffmpeg(paths.filePath)
         .setFfmpegPath(pathToFfmpeg)
         .format('mp3')
-        .audioBitrate(this.state.bitrate)
+        .audioBitrate(bitrate)
         .on('progress', (progress) => {
           // Use same rate limiting as above in function "getVideoAsMp4()" to prevent UI lag.
-          if (!this.rateLimitTriggered) {
-            this.setState({ progress: Math.floor(progress.percent) });
-            this.rateLimitTriggered = true;
+          if (!rateLimitTriggered) {
+            rateLimitTriggered = true;
             setTimeout(() => {
-              this.rateLimitTriggered = false;
+              rateLimitTriggered = false;
             }, 800);
           }
         })
@@ -134,40 +126,42 @@ class AppContainer extends Component {
         )
         .on('end', () => {
           // After the mp3 is wrote to the disk we set the progress to 99% the last 1% is the removal of the temp file.
-          this.setState({ progress: 99 });
+          setProgress(99);
           resolve();
         })
         .run();
     });
   }
 
-  async startDownload(url) {
+  const startDownload = async (url) => {
     console.log('startDownload');
     // Reset state for each download/conversion
-    this.setState({
-      progress: 0,
-      showProgressBar: true,
-      progressMessage: '...',
-    });
+    // ipcRenderer.send('download')
+    const locationSet = changeOutputFolder();
+    if (!locationSet) return;
+
+    setProgress(0);
+    setShowProgress(true);
+    setProgressMessage('...');
 
     try {
       // Tell the user we are getting the video info, and call the function to do so.
-      this.setState({ progressMessage: 'Fetching video info...' });
+      setProgressMessage('Fetching video info...');
       const id = ytdl.getURLVideoID(url);
       const info = await ytdl.getInfo(url);
 
       // Given the id of the video, the path in which to store the output, and the video title
       // download the video as an audio only mp4 and write it to a temp file then return
       // the full path for the tmp file, the path in which its stored, and the title of the desired output.
-      const paths = await this.getVideoAsMp4(
+      const paths = await getVideoAsMp4(
         url,
-        this.state.userDownloadsFolder,
+        folder,
         info.title
       );
 
       // Pass the returned paths and info into the function which will convert the mp4 tmp file into
       // the desired output mp3 file.
-      await this.convertMp4ToMp3(paths);
+      await convertMp4ToMp3(paths);
 
       // Remove the temp mp4 file.
       fs.unlinkSync(paths.filePath);
@@ -176,65 +170,63 @@ class AppContainer extends Component {
       await (() => {
         return new Promise((resolve, reject) => {
           setTimeout(() => {
-            this.setState({ progress: 100 });
+            setProgress(100);
             resolve();
           }, 900);
         });
       });
 
       // Signal that the download and conversion have completed and we need to tell the user about it and then reset.
-      this.downloadFinished();
+      downloadFinished();
     } catch (e) {
       console.error(e);
     }
   }
 
-  downloadFinished() {
+  const downloadFinished = () => {
     // Make sure progress bar is at 100% and tell the user we have completed the task successfully.
-    this.setState({
-      progress: 100,
-      progressMessage: 'Conversion successful!',
-    });
+    setProgress(100);
+    setProgressMessage('Conversion successful!');
 
     // Reset the progress bar to the LinkInput
-    setTimeout(
-      () =>
-        this.setState({
-          showProgressBar: false,
-        }),
-      2000
-    );
+    setTimeout(() => setShowProgress(false), 2000);
   }
 
-  changeOutputFolder() {
+  const changeOutputFolder = () => {
     // Create an electron open dialog for selecting folders, this will take into account platform.
     let fileSelector = remote.dialog.showOpenDialog({
-      defaultPath: `${this.state.userDownloadsFolder}`,
+      defaultPath: `${folder}`,
       properties: ['openDirectory'],
       title: 'Select folder to store files.',
     });
+
+    if(!fileSelector) return;
 
     // If a folder was selected and not just closed, set the localStorage value to that path and adjust the state.
     if (fileSelector) {
       let pathToStore = fileSelector[0];
       localStorage.setItem('userSelectedFolder', pathToStore);
-      this.setState({ userDownloadsFolder: pathToStore });
+      setFolder(pathToStore);
+      return true;
     }
   }
 
-  render() {
-    const { url } = this.state;
-    if (this.state.showProgressBar) {
-      return (
-        <ProgressBar
-          progress={this.state.progress}
-          messageText={this.state.progressMessage}
-        />
-      );
-    } else {
-      return <LinkInput url={url} updateUrl={this.updateUrl} startDownload={this.startDownload} />;
-    }
+  if (showProgress) {
+    return (
+      <ProgressBar
+        progress={progress}
+        messageText={progressMessage}
+      />
+    );
   }
+
+  return (
+    <LinkInput
+      url={url}
+      updateUrl={(e) => setUrl(e.target.value)}
+      startDownload={startDownload}
+    />
+  );
 }
 
 export default AppContainer;
